add all class and modify

+++ src/main/java/com/notimplement/happygear/service/RoleService.java
@@ -3,5 +3,5 @@ package com.notimplement.happygear.service;
 import com.notimplement.happygear.entities.Role;
 
 public interface RoleService {
-    Role getRoleById(int id);
+    Role getRoleById(long id);
 }
